Validate lock file before unlock We need to validate if we have the lock before unlink the file, otherwise we will end up with some issues. Its not a problem of the `delete locks[path]` but it can be a real problem if we dont have the lock, but the lockfile exists and is held by another app, we will try to delete this lockfile without give it any chance to say goodbye.

+++ lockfile.js
@@ -56,16 +56,24 @@ if (/^v0\.[0-8]\./.test(process.version)) {
 
 exports.unlock = function (path, cb) {
   debug('unlock', path)
-  // best-effort.  unlocking an already-unlocked lock is a noop
-  delete locks[path]
-  fs.unlink(path, function (unlinkEr) { cb() })
+
+  if (locks[path]) {
+    delete locks[path]
+    fs.unlink(path, function (unlinkEr) { cb() })
+  } else {
+    cb("No lock found for: " + path)
+  }
 }
 
 exports.unlockSync = function (path) {
   debug('unlockSync', path)
-  // best-effort.  unlocking an already-unlocked lock is a noop
-  try { fs.unlinkSync(path) } catch (er) {}
-  delete locks[path]
+
+  if (locks[path]) {
+    delete locks[path]
+    fs.unlinkSync(path)
+  } else {
+    throw new Error("No lock found for: " + path)
+  }
 }
